# Make default top_k calculation faster