Running `python bench_simple_vector_store.py` 

Before
```
Benchmarking SimpleVectorStore
---------------------------
Adding 10 vectors took 1.3828277587890625e-05 seconds
Querying store of 10 vectors with default mode took 0.0025229454040527344 seconds
Adding 50 vectors took 5.4836273193359375e-05 seconds
Querying store of 50 vectors with default mode took 0.01211094856262207 seconds
Adding 100 vectors took 9.274482727050781e-05 seconds
Querying store of 100 vectors with default mode took 0.02292323112487793 seconds
Adding 500 vectors took 0.00044608116149902344 seconds
Querying store of 500 vectors with default mode took 0.11435413360595703 seconds
Adding 1000 vectors took 0.0010182857513427734 seconds
Querying store of 1000 vectors with default mode took 0.2347888946533203 seconds
Adding 10000 vectors took 0.008826971054077148 seconds
Querying store of 10000 vectors with default mode took 2.2964768409729004 seconds
```

After
```
Benchmarking SimpleVectorStore
---------------------------
Adding 10 vectors took 1.5974044799804688e-05 seconds
Querying store of 10 vectors with default mode took 0.0007221698760986328 seconds
Adding 50 vectors took 4.506111145019531e-05 seconds
Querying store of 50 vectors with default mode took 0.0030128955841064453 seconds
Adding 100 vectors took 8.7738037109375e-05 seconds
Querying store of 100 vectors with default mode took 0.005853176116943359 seconds
Adding 500 vectors took 0.0006880760192871094 seconds
Querying store of 500 vectors with default mode took 0.030952930450439453 seconds
Adding 1000 vectors took 0.0008981227874755859 seconds
Querying store of 1000 vectors with default mode took 0.05759096145629883 seconds
Adding 10000 vectors took 0.00973820686340332 seconds
Querying store of 10000 vectors with default mode took 0.5656261444091797 seconds
```